add poly parameter in diso() function
a precursor-dereived metabolite might be the modification of the dimer , trimer or even polymer of the precursor. For instance, [2Try + x + H]+. In this situation, the mass difference of this labeled and unlabed metabolite will be 2 times of this labeled precursor. Poly parameter is therefore used to catch this kind of metabolotes.